requirements.txt
black==21.9b0
click==8.0.3
mypy-extensions==0.4.3
numpy==1.26.4
pathspec==0.9.0
Pillow==9.3.0
platformdirs==2.4.0
regex==2021.10.8
tomli==1.2.1
typing-extensions==3.10.0.2
